[STORM-1714] refactored common logic into BaseStatefulBoltExecutor
